chore(hygiene): fixtures match_registry alignees, ratchets DDL et titleseams durcis (etape 5)
- Aide de `--gamertag` de backfill-killsource alignee sur le comportement reel (le joueur
  nomme est celui DONT les films sont traites ; les jetons viennent du pool), message
  d'erreur de --online idem.
- Deux fixtures recopiaient encore `team_{0,1}_score SMALLINT` : alignees sur INTEGER. La DDL
  legacy VOLONTAIRE du test d'elargissement porte desormais le marqueur
  `match_registry-ddl: legacy — <raison>`.
- Nouveau ratchet TestMatchRegistryFixturesDeTestAlignees : toute fixture non marquee declare,
  pour ses colonnes communes avec la production, les MEMES types (68 fixtures examinees,
  parseur tolerant a la mise en forme). Premiere mesure : 45 divergences preexistantes dans 25
  fichiers, toutes sur les horodatages / backfill_completed / player_count — hors perimetre,
  GELEES dans une carte datee qui ne peut que retrecir (une entree realignee rougit aussi).
- Ratchet titleseams : critere = dependances TRANSITIVES (un seul `go list`, t.Skip sans go au
  PATH), correspondance exacte sur les deux paquets porteurs d'un fail-loud (internal/sync et
  internal/sync/skill). Les 10 binaires annonces recoivent `titleseams.RegisterAll("")` ; 42
  binaires comptes, tous cables.
- COMMANDS.md FR + EN : note d'exploitation (base partagee tenue en ecriture + migrations
  appliquees -> serveur arrete ; ne pas faire tourner les jetons du parc pendant qu'un serveur
  tourne).

Gate G5 : archlint + migrations du titre → 0 ; complement -tags=integration -p 1 sur
internal/migration et internal/sync (les fixtures touchees y sont) → 0 ; go build ./cmd/... → 0.
Trois mutations verifiees (DDL non marquee, entree gelee perimee, cablage retire d'un binaire
transitif) : chacune fait rougir son ratchet.

+++ apps/go-api/internal/sync/art_rebuild_regression_test.go
@@ -107,8 +107,8 @@ func seedFullSchemaForARTRebuild(t *testing.T, db *sql.DB) {
 			is_firefight BOOLEAN DEFAULT FALSE,
 			duration_seconds INTEGER,
 			playable_duration_seconds INTEGER,
-			team_0_score SMALLINT,
-			team_1_score SMALLINT,
+			team_0_score INTEGER,
+			team_1_score INTEGER,
 			team_0_ps_score INTEGER,
 			team_1_ps_score INTEGER,
 			player_count INTEGER
